Moved convention search components inside search directory
